# book-network/src/main/java/com/verytroll/book_network/feedback/FeedbackResponse.java
package com.verytroll.book_network.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {
    private Double note;
    private String comment;
    private boolean ownFeedback;
}